Implement feature request #4237: see new sample: area/multi-axis-with-seriesname-arrays.xml
- also use this facility to simplify the
specification of series inclusion in stacked charts, especially when
combined with other series.

See stacked-column-with-line.xml
and c.f. stacked-column-with-line-new.xml

The new facility should be backward compatible with the existing
style (both described below to the best of my knowledge), via code
that converts the old data structure to the new one.

The current style of mapping multiple series to a y axis is to
include one yaxis config per series but set each yaxis seriesName to
the same series name and then set show: false on all axes that are not
required.

While not the purpose of #4237, it offers a more intuitive alternative
that explicitly configures yaxis elements with the series that will be
referenced to them (seriesName: []). This only requires including the
yaxis elements that will be seen on the chart.

Old way:
  yax:	ser
  0:	0
  1:	1
  2:	1
  3:	1
  4:	1

Axes 0..4 are all scaled and all will be rendered unless the axes are
show: false. If the chart is stacked, it's assumed that series 1..4 are
the contributing series.

New way:
  yax:	ser
  0:	[0]
  1:	[1,2,3,4]

If the chart is stacked (a global chart setting), it must be assumed
that any axis with multiple series is stacked, presumably as separate
sets (not yet implemented).
 

Fix stacked chart y axis scaling: min and max were computed as the
sum of the series min and max, not the min and max od the summed
datapoints in each series.

Note: it is left to the creativity of the user to visually disambiguate
which plot is associated with which Y axis.

Additional checks for undefined elements in CoreUtils.extendArrayProps().

Remove the duplicate outlier in boxplot-scatter.xml.