Removal of Scales.autoScaleY() and all calls to it: Options.chart.zoom.autoScaleYaxis had no effect, now fixed: Possibly since release 3.46.0 as a result of changes related to brush charts, Y axes are always scaled when zoomed regardless of autoScaleYaxis, either in the options.chart.zoom or options.chart.brush. The correct behaviour has been restored but is controlled within Range.getMinYMaxY().
A note for debugging:
This routine is called numerous times during chart preparation,
including before all charts are fully populated with state relating to
the brush option. The crucial pass through this function is, I believe,
as a result of the event trigger to update the target chart with zoom
information, when it is then possible to determine the required
brushSource state.

Scales.autoScaleY() and calls to it have been removed completely as
they are no longer needed, and was duplicating the role of
Range.getMinYMaxY().

This change exposed the disableDefault in passive event listener common
issue, which was resolved by setting "passive: false" in svg.js.