Additional fixes for indexing errors using yaxis-series mappings after seriesName-as-an-array feature implementation.
Add additional checks to manage old and new modes.

Unit tests targeting seriesName-as-an-array feature have been modified
and expanded to "prove" yaxis-series forward and reverse mapping
correctness.

Context:
Backward compatibility requires that various data structures be
indexable by either the yaxis index or the series index at different
times, which rely on the old one-to-one correspondence [in multi axis
charts]. The new seriesName-as-an-array feature removes this
correspondence and implements forward and reverse maps to bridge
compatibility, nevertheless, every reference into those arrays in the
codebase needs to be discovered and adjusted.