Fix for #4357
Add a missing conditional block designed to ensure the range remains a
multiple of stepSize when min is defined. Was present in the max case
but not the min case.

Record the absolute final tick number used in the grid alignment of
following axes, that is, after pruning to maxTicks.