Added support for the Angular compiler.
The Angular compiler looks for the main field using a few fields in the package.json. The order it uses is:
- es2020
- es2015
- browser
- module
- main

The first entry the Angular compiler would find before this change was the 'main' entry point. This is however compiled as `common js` and Angular is looking for an `esm` so it would complain (it would still work though). Adding `es2020` as an entry point fixes this problem.