Issue 4704 gl.seriesX[0].length() was effectively used to size a number of temporary arrays in Scales.scaleMultipleYAxes() but it appears that gl.seriesX is not populated if tickPlacement: 'between' is configured, so this initialization now uses gl.datapoints instead.
Removed a duplicate version of a function that was left behind during
previous refactoring:
Scales.setSeriesYAxisMappings()
CoreUtils.setSeriesYAxisMappings()
Both were identical except for a loop at the end of
CoreUtils.setSeriesYAxisMappings() which set default series group
names. Therefore removing Scales.setSeriesYAxisMappings() and
substituting the call to it should have no net effect. This appears to
be the case from testing.