Remove use of global `ApexCharts` class
The `setupBrushHandler` method of the `Core` class relies on the
`ApexCharts` class' static `getChartByID` method. It calls this
static method on the `ApexCharts` class directly without
importing it currently.

This is fine if consuming `apexcharts` via
CDN / script tag as that places the exported `ApexCharts` class in
the global scope, but it becomes an issue when users want to
consume `apexcharts` via a package manager.

Update the `setupBrushHandler` method to call `ApexCharts.getChartByID`
the `Core` class' `ctx` field instead.