range: Fix getting decimal part of numbers close to 0
Previously, isFloat was being called on val before noExponents. This
meant that if a value was particularly close to zero, it might pass the
isFloat check but then be stripped down to just 0 by noExponents, which
meant that getting the second element after splitting around the decimal
point would fail.